requirements.txt
Flask==2.0.1
werkzeug==2.0.1
python-dotenv==1.0.0
pyodbc
flask_httpauth
Flask-SQLAlchemy
Flask-Migrate
Flask-WTF
requests
azure-search-documents==11.4.0  # Azure Cognitive Search 用パッケージ
azure-core>=1.28.0,<2.0.0  # Azure SDK のコアライブラリ
openai==0.27.0  # OpenAI API用パッケージ
pypdf==3.8.1  # PDF操作用パッケージ
